telemetry resource: update etcd path to use policy name (#1901)
### Description of change

Prevent clash in telemetry collector etcd paths across policies

+++ pkg/policies/controlplane/policy.go
@@ -120,7 +120,7 @@ func compilePolicyWrapper(wrapperMessage *policysyncv1.PolicyWrapper, registry s
 			// Initialize classifiers
 			classifiers := flowControl.GetClassifiers()
 			for index, classifierProto := range classifiers {
-				classifierOption, err := classifier.NewClassifierOptions(int64(index), classifierProto, policy)
+				classifierOption, err := classifier.NewClassifierOptions(index, classifierProto, policy)
 				if err != nil {
 					return nil, nil, nil, err
 				}
